+++ src/main/java/de/unistuttgart/iste/gits/template/persistence/entity/package-info.java
@@ -1,5 +1,5 @@
 /**
- * This package contains the DAOs (database access objects) for the persistence layer.
+ * This package contains the entities for the persistence layer.
  * <p>
  * Information on how to define Entities can be found in <a href=https://www.baeldung.com/jpa-entities>this tutorial</a>
  * or <a href=https://docs.oracle.com/javaee/7/tutorial/persistence-intro.htm#BNBPY>this tutorial</a>
@@ -10,4 +10,4 @@
  * To avoid having to implement getters, setters, builders, etc. we use <a href=https://projectlombok.org/>Lombok</a>.
  * The Lombok annotations (e.g. {@code @Data}) should be used in the entities.
  */
-package de.unistuttgart.iste.gits.template.persistence.dao;
+package de.unistuttgart.iste.gits.template.persistence.entity;
