+++ src/main/java/de/unistuttgart/iste/gits/template/service/TemplateService.java
@@ -1,7 +1,7 @@
 package de.unistuttgart.iste.gits.template.service;
 
 import de.unistuttgart.iste.gits.generated.dto.Template;
-import de.unistuttgart.iste.gits.template.persistence.dao.TemplateEntity;
+import de.unistuttgart.iste.gits.template.persistence.entity.TemplateEntity;
 import de.unistuttgart.iste.gits.template.persistence.mapper.TemplateMapper;
 import de.unistuttgart.iste.gits.template.persistence.repository.TemplateRepository;
 import lombok.RequiredArgsConstructor;
